requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.115
botocore==1.29.115
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
Django==4.1.7
django-environ==0.10.0
django-geolocation==0.2
django-ratelimit==4.0.0
django-redis==5.2.0
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
geographiclib==2.0
geopy==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.5
openapi-codec==1.3.2
packaging==23.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
service-identity==21.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
zope.interface==6.0
